Backend/requirement.txt
asgiref==3.2.7
automium==0.2.6
automium-web==0.1.1
click==7.1.2
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.19.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Werkzeug==1.0.1
